Add initial release script
This script can be run by doing `npm run-script release` only on release
branches. Additionally, the second part of the release branch name
should be the current version number. So, if the current version is
"0.0.1", the release branch name should be "release/0.0.1". Otherwise
this script will not work.

This script copies the files from build folder to the themes folder
currently, and warns for files that have been replaced or removed in the process.